feat: make event be in the index page and add default 404 page
