+++ src/app/layout.tsx
@@ -1,29 +1,6 @@
 'use client'
-import { Inter } from 'next/font/google';
-import NavBar from './components/molecules/nav-bar';
-import { useTitle, TitleProvider } from './providers/TitleContext';
-import Head from 'next/head';
-import "./globals.css";
-import { ListingHeader } from './components/molecules/listing-header';
+import { TitleProvider } from './providers/TitleContext';
 
-
-const inter = Inter({ subsets: ['latin'] });
-
-const LayoutContent: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-  const { title } = useTitle();
-  return (
-    <div className={`flex flex-col md:flex-row overflow-hidden h-screen w-screen ${inter.className}`}>
-      <NavBar />
-      <div className="flex flex-col pt-2 pb-8 px-6 bg-secondary overflow-hidden w-full h-full gap-2 md:gap-6">
-        <h1 className="text-2xl text-gray-700">{title}</h1>
-        <header>
-          <ListingHeader/>
-        </header>
-        <main className="flex flex-grow overflow-hidden p-2 md:p-4 bg-white">{children}</main>
-      </div>
-    </div>
-  );
-};
 
 export default function RootLayout({
   children,
@@ -34,7 +11,7 @@ export default function RootLayout({
     <TitleProvider>
       <html lang="en">
         <body suppressHydrationWarning={true}>
-          <LayoutContent>{children}</LayoutContent>
+          {children}
         </body>
       </html>
     </TitleProvider>
